# Server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Minus Trash API",
    "version": "1.0.0",
    "description": "Documentazione API per Minus Trash"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
      "User": {
        "title": "User",
        "properties": {
          "username": {
            "type": "string"
          },
          "fullName": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "surname": {
                "type": "string"
              }
            }
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "cittadino",
              "operatore_comunale",
              "amministratore"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "blockedAt": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "authMethods": {
            "type": "object",
            "properties": {
              "local": {
                "type": "boolean"
              },
              "google": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          },
          "lastLogin": {
            "type": "string"
          },
          "activeSessions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "device": {
                  "type": "string"
                },
                "ip": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string"
                }
              }
            }
          },
          "passwordChangedAt": {
            "type": "string"
          },
          "_id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "type": "object",
        "example": {
          "email": "test@test.com",
          "password": "Securepass123!",
          "fullName": {
            "name": "Test",
            "surname": "Test"
          },
          "username": "test",
          "role": "cittadino"
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Registra un nuovo utente",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utente registrato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Dati non validi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Utente già registrato"
                    },
                    "errors": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "example": "Email già registrata"
                          }
                        },
                        "username": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "example": "Username già registrato"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login utente (richiede Basic Auth)",
        "security": [
          {
            "BasicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Login effettuato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenziali non valide"
          }
        }
      }
    },
    "/api/auth/login-test": {
      "post": {
        "summary": "Login utente (test senza Basic Auth)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login effettuato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenziali non valide"
          }
        }
      }
    },
...
    "/api/auth/googleOAuth/callback": {
      "get": {
        "summary": "Callback dopo autenticazione Google",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Autenticazione completata"
          },
          "401": {
            "description": "Autenticazione fallita"
          }
        }
      }
    },
    "/api/auth/profile_update": {
      "post": {
        "summary": "Aggiorna il profilo utente",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "surname": {
                        "type": "string"
                      }
                    }
                  },
                  "username": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profilo aggiornato con successo"
          }
        }
      }
    },
    "/api/auth/change_password": {
      "post": {
        "summary": "Cambia password utente",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password cambiata con successo"
          },
          "400": {
            "description": "Errore nel cambio password"
          }
        }
      }
    },
    "/api/auth/users": {
      "get": {
        "summary": "Ottieni tutti gli utenti (Solo amministratore)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista di tutti gli utenti recuperata con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non autorizzato - Token mancante o non valido"
          },
          "403": {
            "description": "Vietato - L'utente non ha privilegi di amministratore"
          }
        }
      },
      "post": {
        "summary": "Crea un nuovo utente (Solo amministratore)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utente creato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Dati non validi"
          },
          "401": {
            "description": "Non autorizzato - Token mancante o non valido"
          },
          "403": {
            "description": "Vietato - L'utente non ha privilegi di amministratore"
          }
        }
      }
    },
    "/api/auth/users/{userId}": {
      "put": {
        "summary": "Aggiorna un utente tramite ID (Solo amministratore)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "L'ID dell'utente da aggiornare"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Utente aggiornato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Dati non validi"
          },
          "401": {
            "description": "Non autorizzato - Token mancante o non valido"
          },
          "403": {
            "description": "Vietato - L'utente non ha privilegi di amministratore"
          },
          "404": {
            "description": "Utente non trovato"
          }
        }
      },
      "delete": {
        "summary": "Elimina un utente tramite ID (Solo amministratore)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "L'ID dell'utente da eliminare"
          }
        ],
        "responses": {
          "200": {
            "description": "Utente eliminato con successo"
          },
          "401": {
            "description": "Non autorizzato - Token mancante o non valido"
          },
          "403": {
            "description": "Vietato - L'utente non ha privilegi di amministratore"
          },
          "404": {
            "description": "Utente non trovato"
          }
        }
      }
    },
...
  }
}